requirements.txt
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djoser==2.2.0
drf-base64==2.0
flake8==6.0.0
gunicorn==20.1.0
Pillow==10.0.0
psycopg2==2.9.6
django-colorfield==0.9.0
python-dotenv==1.0.0
